requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
idna==2.10
praw==7.1.0
prawcore==1.5.0
pycparser==2.20
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
tinydb==4.1.1
tornado==6.0.4
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
